requirements.txt
requests
beautifulsoup4
Pillow
pypdf
python-dotenv
